Hormonize packages for FTP & SFTP modules
