+++ spring-integration-ftp/src/main/resources/org/springframework/integration/ftp/config/spring-integration-ftp.xsd
@@ -235,7 +235,7 @@
 		<xsd:annotation>
 			<xsd:documentation>
 				Configures a Consumer Endpoint for the
-				'org.springframework.integration.ftp.gateway.FtpOutboundGateway'
+				'org.springframework.integration.ftp.outbound.FtpOutboundGateway'
 				used to issue FTP commands.
 			</xsd:documentation>
 		</xsd:annotation>
