+++ src/reference/antora/modules/ROOT/pages/ftp/session-callback.adoc
@@ -20,5 +20,5 @@ Another example might be to pre- or post-process the file data being sent or ret
 When using XML configuration, the `<int-ftp:outbound-gateway/>` provides a `session-callback` attribute to let you specify the `MessageSessionCallback` bean name.
 
 NOTE: The `session-callback` is mutually exclusive with the `command` and `expression` attributes.
-When configuring with Java, different constructors are available in the https://docs.spring.io/spring-integration/api/org/springframework/integration/ftp/gateway/FtpOutboundGateway.html[`FtpOutboundGateway`] class.
+When configuring with Java, different constructors are available in the https://docs.spring.io/spring-integration/api/org/springframework/integration/ftp/outbound/FtpOutboundGateway.html[`FtpOutboundGateway`] class.
